pyproject.toml
[project]
name = "unified-m"
version = "0.2.0"
description = "Unified Marketing Measurement: one incrementality truth from MMM + attribution + lift tests"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Unified-M Contributors" }]

dependencies = [
    # Core (always required)
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",

    # Validation & config
    "pydantic>=2.6.0",
    "pyyaml>=6.0.0",

    # API (also serves the built React UI in production)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",

    # CLI & logging
    "typer>=0.9.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
# Heavy model backends (install only what you need)
pymc = [
    "pymc>=5.10.0",
    "pymc-marketing>=0.9.0",
    "arviz>=0.18.0",
]
# Data connectors
connectors = [
    "polars>=1.0.0",
    "pyarrow>=15.0.0",
    "duckdb>=1.0.0",
]
# Full install
all = [
    "unified-m[pymc,connectors]",
    "pandera>=0.20.0",
    "altair>=5.2.0",
    "python-dotenv>=1.0.0",
]
# Development
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
unified-m = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src"
